refactor: DB 전반적인 예외 메세지에 대한 의미가 드러나도록 DatabaseError로 수정
